Create a common presenter
Both WorldLocationPresenter and WorldwideOrganisationPresenter have been
using the exact same code, this commit aims to reduce duplication by
introducing a common object that both classes inherit from.

This way still gives freedom for both classes to part ways if needed in
the future.
